feature(events): Allow class name as event/hook handler
Fixes #8077
