+++ engine/classes/Elgg/PluginHooksService.php
@@ -25,11 +25,29 @@ class PluginHooksService extends \Elgg\HooksRegistrationService {
 		$hooks = $this->getOrderedHandlers($hook, $type);
 		
 		foreach ($hooks as $callback) {
+
+			$callback_copy = $callback;
+
+			if (is_string($callback)
+					&& false === strpos($callback, '::')
+					&& !function_exists($callback)
+					&& class_exists($callback)) {
+
+				$class = ltrim($callback, '\\');
+
+				$cached = _elgg_services()->config->get("object:$class");
+				if (!$cached) {
+					$cached = new $class();
+					_elgg_services()->config->set("object:$class", $cached);
+				}
+				$callback = $cached;
+			}
+
 			if (!is_callable($callback)) {
 				if ($this->logger) {
 					$inspector = new Inspector();
 					$this->logger->warn("handler for plugin hook [$hook, $type] is not callable: "
-										. $inspector->describeCallable($callback));
+										. $inspector->describeCallable($callback_copy));
 				}
 				continue;
 			}
